refactor: break 2 circular import dependency cycles
Cycle 1: ui/action.tsx -> action_link.tsx -> action_helpers.ts -> action.tsx
- Extract ActionVariant, ActionSize types and actionClassName() into new
  action_base.ts module (lowest-level, no cycle imports)
- Update action_helpers.ts to import from action_base.ts
- action.tsx re-exports from action_base.ts

Cycle 2: health_bar.tsx <-> health_status.tsx
- Extract severityVariant() into new health_utils.ts module
- Update health_bar.tsx, health_status.tsx, and store_card/index.tsx
  to all import from health_utils.ts
