+++ app/frontend/pages/admin/dashboard/store_card/health_bar.tsx
@@ -2,10 +2,7 @@ import type { AdminStoreSummary } from "@/types/inertia";
 
 import HealthStatus from "./health_status";
 import { StoreHealthMetrics } from "./store_health_metrics";
-
-export function severityVariant(severity: string) {
-  return severity as "good" | "working" | "warning" | "danger" | "neutral";
-}
+import { severityVariant } from "./health_utils";
 
 export default function StoreHealthBar({ store }: { store: AdminStoreSummary }) {
   return (
