+++ app/frontend/pages/admin/dashboard/store_card/health_utils.ts
@@ -1,0 +1,3 @@
+export function severityVariant(severity: string) {
+  return severity as "good" | "working" | "warning" | "danger" | "neutral";
+}
